+++ src/multipac_testbench/instruments/step_constant.py
@@ -12,6 +12,7 @@ import pandas as pd
 from multipac_testbench.instruments.power import Sync
 from multipac_testbench.instruments.virtual_instrument import VirtualInstrument
 from multipac_testbench.multipactor_test.helper import parse_header_value
+from multipac_testbench.util.filtering import not_noisy_array_is_growing
 from multipac_testbench.util.types import POST_TREATER_T
 from numpy.typing import NDArray
 
@@ -92,6 +93,9 @@ class StepConstant(VirtualInstrument):
         """
         return []
 
+    def growth_array(self, **kwargs) -> NDArray[np.float64]:
+        return not_noisy_array_is_growing(self.data, **kwargs)
+
 
 class CurrentCalibre(StepConstant):
     """Store the current calibre.
@@ -169,6 +173,7 @@ class PowerSetpoint(StepConstant):
         width: int = 2,
         **kwargs,
     ) -> NDArray[np.bool]:
+        raise ValueError("should not be used i think")
         return super().growth_mask(
             minimum_number_of_points=minimum_number_of_points,
             n_trailing_points_to_check=n_trailing_points_to_check,