+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - M. Vénière correction function was not applied.
+- Power evolution with seesaw-like profile is better handled. In particular,
+  detection of multipactor.
 
 ## [2.0.7] - 2026-06-05